simplify ResponsiveGridFlowLayout code
